refactor: isolate daemon artifact HTTP routes
